requirements.txt
torchmetrics==0.11.4
einops
mne==1.6.0
hydra-core==1.3.2
wandb==0.15.0
termcolor
numpy==1.26